+++ src/components/molecules/Toolbar/index.tsx
@@ -3,11 +3,11 @@ import { Grid, Toolbar as MuiToolbar } from "@material-ui/core";
 import { Add } from "@material-ui/icons";
 import SearchInput from "../../atoms/SearchInput";
 import DialogButton from "../../atoms/DialogButton";
-import { useStyles } from "./style";
+import { useStyles } from "./styles";
 import { useDispatch } from "react-redux";
 import { openAddExpenseModal } from "../../../slices/ExpensesSlice";
 
-const Toolbar = () => {
+const Toolbar: React.FC = () => {
   const classes = useStyles();
   const dispatch = useDispatch();
