+++ src/components/templates/Dialog/index.tsx
@@ -3,8 +3,6 @@ import {
   Dialog as MuiDialog,
   DialogTitle,
   DialogContent,
-  makeStyles,
-  Theme,
   Typography,
 } from "@material-ui/core";
 import Form from "../../organisms/Form";
@@ -15,22 +13,11 @@ import {
   closeAddExpenseModal,
   closeEditExpenseModal,
 } from "../../../slices/ExpensesSlice";
+import { useStyles } from "./styles";
 
 interface DialogProps {
   actionType: "edit" | "add";
 }
-
-const useStyles = makeStyles((theme: Theme) => ({
-  dialogWrapper: {
-    padding: theme.spacing(2),
-    position: "absolute",
-    top: theme.spacing(5),
-  },
-
-  dialogTitle: {
-    padding: "0px",
-  },
-}));
 
 const Dialog: React.FC<DialogProps> = ({ actionType }) => {
   const dispatch = useDispatch();
