+++ src/components/templates/Dialog/styles.tsx
@@ -1,0 +1,13 @@
+import { makeStyles, Theme } from "@material-ui/core";
+
+export const useStyles = makeStyles((theme: Theme) => ({
+  dialogWrapper: {
+    padding: theme.spacing(2),
+    position: "absolute",
+    top: theme.spacing(5),
+  },
+
+  dialogTitle: {
+    padding: "0px",
+  },
+}));
